+++ R/betaSandwich-beta-adf.R
@@ -19,10 +19,17 @@
 #'     \item{lm}{Object of class `lm`.}
 #'     \item{lm_process}{Pre-processed object of class `lm`.}
 #'     \item{type}{Standard error type.}
+#'     \item{gamman}{Asymptotic covariance matrix
+#'       of the sample covariance matrix
+#'       assuming multivariate normality.}
+#'     \item{gammahc}{Asymptotic covariance matrix
+#'       HC correction.}
 #'     \item{gamma}{Asymptotic covariance matrix
 #'       of the sample covariance matrix.}
-#'     \item{acov}{Asymptotic covariance matrix of the standardized slopes.}
-#'     \item{vcov}{Sampling covariance matrix of the standardized slopes.}
+#'     \item{acov}{Asymptotic covariance matrix
+#'       of the standardized slopes.}
+#'     \item{vcov}{Sampling covariance matrix
+#'       of the standardized slopes.}
 #'     \item{est}{Vector of standardized slopes.}
 #'   }
 #'
@@ -31,26 +38,12 @@
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaADF(object)
-#' # Methods ----------------------------------------------------
+#' # Methods -------------------------------------------------------
 #' print(std)
 #' summary(std)
 #' coef(std)
 #' vcov(std)
 #' confint(std, level = 0.95)
-#' ## Differences of standardized regression coefficients -------
-#' out <- dif(std)
-#' print(out)
-#' summary(out)
-#' coef(out)
-#' vcov(out)
-#' confint(out, level = 0.95)
-#' ## Multiple Correlation --------------------------------------
-#' out <- rsq(std)
-#' print(out)
-#' summary(out)
-#' coef(out)
-#' vcov(out)
-#' confint(out, level = 0.95)
 #' @export
 #' @family Beta Sandwich Functions
 #' @keywords betaSandwich
@@ -113,6 +106,11 @@ BetaADF <- function(object) {
     lm = object,
     lm_process = lm_process,
     type = "adf",
+    gamman = .GammaN(
+      sigmacap = lm_process$sigmacap,
+      pinv_of_dcap = .PInvDmat(.DMat(lm_process$k))
+    ),
+    gammahc = NULL,
     gamma = gammacap_adf,
     acov = chol2inv(chol(acov)),
     vcov = vcov,