+++ R/betaSandwich-beta-hc.R
@@ -4,20 +4,25 @@
 #' @author Ivan Jacob Agaloos Pesigan
 #'
 #' @return Returns an object of class `betasandwich`
-#' which is a list with the following elements:
-#' \describe{
-#'   \item{call}{Function call.}
-#'   \item{lm}{Object of class `lm`.}
-#'   \item{lm_process}{Pre-processed object of class `lm`.}
-#'   \item{type}{Standard error type.}
-#'   \item{gamman}{Asymptotic covariance matrix of the sample covariance matrix
-#'     assuming multivariate normal distribution.}
-#'   \item{gammahc}{Asymptotic covariance matrix HC adjustment.}
-#'   \item{gamma}{Asymptotic covariance matrix of the sample covariance matrix.}
-#'   \item{acov}{Asymptotic covariance matrix of the standardized slopes.}
-#'   \item{vcov}{Sampling covariance matrix of the standardized slopes.}
-#'   \item{est}{Vector of standardized slopes.}
-#' }
+#'   which is a list with the following elements:
+#'   \describe{
+#'     \item{call}{Function call.}
+#'     \item{lm}{Object of class `lm`.}
+#'     \item{lm_process}{Pre-processed object of class `lm`.}
+#'     \item{type}{Standard error type.}
+#'     \item{gamman}{Asymptotic covariance matrix
+#'       of the sample covariance matrix
+#'       assuming multivariate normality.}
+#'     \item{gammahc}{Asymptotic covariance matrix
+#'       HC correction.}
+#'     \item{gamma}{Asymptotic covariance matrix
+#'       of the sample covariance matrix.}
+#'     \item{acov}{Asymptotic covariance matrix
+#'       of the standardized slopes.}
+#'     \item{vcov}{Sampling covariance matrix
+#'       of the standardized slopes.}
+#'     \item{est}{Vector of standardized slopes.}
+#'   }
 #'
 #' @param object Object of class `lm`.
 #' @param type Character string.
@@ -47,26 +52,12 @@
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaHC(object)
-#' # Methods ----------------------------------------------------
+#' # Methods -------------------------------------------------------
 #' print(std)
 #' summary(std)
 #' coef(std)
 #' vcov(std)
 #' confint(std, level = 0.95)
-#' ## Differences of standardized regression coefficients -------
-#' out <- dif(std)
-#' print(out)
-#' summary(out)
-#' coef(out)
-#' vcov(out)
-#' confint(out, level = 0.95)
-#' ## Multiple Correlation --------------------------------------
-#' out <- rsq(std)
-#' print(out)
-#' summary(out)
-#' coef(out)
-#' vcov(out)
-#' confint(out, level = 0.95)
 #' @export
 #' @family Beta Sandwich Functions
 #' @keywords betaSandwich