+++ R/betaSandwich-methods-diff-beta-sandwich.R
@@ -1,4 +1,4 @@
-#' Print Method for an Object of Class `difbetasandwich`
+#' Print Method for an Object of Class `diffbetasandwich`
 #'
 #' @author Ivan Jacob Agaloos Pesigan
 #'
@@ -10,7 +10,7 @@
 #'   and
 #'   confidence intervals.
 #'
-#' @param x Object of class `difbetasandwich`.
+#' @param x Object of class `diffbetasandwich`.
 #' @param ... additional arguments.
 #' @param alpha Significance level.
 #' @param digits Digits to print.
@@ -18,14 +18,14 @@
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaHC(object)
-#' out <- dif(std)
-#' print(out)
+#' diff <- DiffBetaSandwich(std)
+#' print(diff)
 #' @export
 #' @keywords methods
-print.difbetasandwich <- function(x,
-                                  alpha = c(0.05, 0.01, 0.001),
-                                  digits = 4,
-                                  ...) {
+print.diffbetasandwich <- function(x,
+                                   alpha = c(0.05, 0.01, 0.001),
+                                   digits = 4,
+                                   ...) {
   cat(
     "Difference between standardized regression coefficients with",
     toupper(x$fit$type),
@@ -42,7 +42,7 @@ print.difbetasandwich <- function(x,
   )
 }
 
-#' Summary Method for an Object of Class `difbetasandwich`
+#' Summary Method for an Object of Class `diffbetasandwich`
 #'
 #' @author Ivan Jacob Agaloos Pesigan
 #'
@@ -54,7 +54,7 @@ print.difbetasandwich <- function(x,
 #'   and
 #'   confidence intervals.
 #'
-#' @param object Object of class `difbetasandwich`.
+#' @param object Object of class `diffbetasandwich`.
 #' @param ... additional arguments.
 #' @param alpha Significance level.
 #' @param digits Digits to print.
@@ -62,14 +62,14 @@ print.difbetasandwich <- function(x,
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaHC(object)
-#' out <- dif(std)
-#' summary(out)
+#' diff <- DiffBetaSandwich(std)
+#' summary(diff)
 #' @export
 #' @keywords methods
-summary.difbetasandwich <- function(object,
-                                    alpha = c(0.05, 0.01, 0.001),
-                                    digits = 4,
-                                    ...) {
+summary.diffbetasandwich <- function(object,
+                                     alpha = c(0.05, 0.01, 0.001),
+                                     digits = 4,
+                                     ...) {
   cat(
     "Difference between standardized regression coefficients with",
     toupper(object$fit$type),
@@ -95,18 +95,18 @@ summary.difbetasandwich <- function(object,
 #'   variance-covariance matrix
 #'   of differences of standardized regression slopes.
 #'
-#' @param object Object of class `difbetasandwich`.
+#' @param object Object of class `diffbetasandwich`.
 #' @param ... additional arguments.
 #'
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaHC(object)
-#' out <- dif(std)
-#' vcov(out)
+#' diff <- DiffBetaSandwich(std)
+#' vcov(diff)
 #' @export
 #' @keywords methods
-vcov.difbetasandwich <- function(object,
-                                 ...) {
+vcov.diffbetasandwich <- function(object,
+                                  ...) {
   return(
     object$vcov
   )
@@ -118,18 +118,18 @@ vcov.difbetasandwich <- function(object,
 #'
 #' @return Returns a vector of differences of standardized regression slopes.
 #'
-#' @param object Object of class `difbetasandwich`.
+#' @param object Object of class `diffbetasandwich`.
 #' @param ... additional arguments.
 #'
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaHC(object)
-#' out <- dif(std)
-#' coef(out)
+#' diff <- DiffBetaSandwich(std)
+#' coef(diff)
 #' @export
 #' @keywords methods
-coef.difbetasandwich <- function(object,
-                                 ...) {
+coef.diffbetasandwich <- function(object,
+                                  ...) {
   return(
     object$est
   )
@@ -142,7 +142,7 @@ coef.difbetasandwich <- function(object,
 #'
 #' @return Returns a matrix of confidence intervals.
 #'
-#' @param object Object of class `difbetasandwich`.
+#' @param object Object of class `diffbetasandwich`.
 #' @param ... additional arguments.
 #' @param parm a specification of which parameters
 #'   are to be given confidence intervals,
@@ -153,14 +153,14 @@ coef.difbetasandwich <- function(object,
 #' @examples
 #' object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
 #' std <- BetaHC(object)
-#' out <- dif(std)
-#' confint(out, level = 0.95)
+#' diff <- DiffBetaSandwich(std)
+#' confint(diff, level = 0.95)
 #' @export
 #' @keywords methods
-confint.difbetasandwich <- function(object,
-                                    parm = NULL,
-                                    level = 0.95,
-                                    ...) {
+confint.diffbetasandwich <- function(object,
+                                     parm = NULL,
+                                     level = 0.95,
+                                     ...) {
   if (is.null(parm)) {
     parm <- seq_len(
       length(object$est)